fix: remove envs from __mocks__ and PUBLIC_URL
